# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "English Brain API",
    "description": "English Brain Backend - Orchestrator for STT, Ollama LLM, TTS and FSRS Spaced Repetition",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/questions": {
      "get": {
        "tags": [
          "questions"
        ],
        "summary": "List Questions",
        "operationId": "list_questions_api_questions_get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category (hr, tech, vocab)",
              "title": "Category"
            },
            "description": "Filter by category (hr, tech, vocab)"
          },
          {
            "name": "difficulty",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by difficulty (junior, mid, senior)",
              "title": "Difficulty"
            },
            "description": "Filter by difficulty (junior, mid, senior)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionOut"
                  },
                  "title": "Response List Questions Api Questions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/questions/{question_id}": {
      "get": {
        "tags": [
          "questions"
        ],
        "summary": "Get Question",
        "operationId": "get_question_api_questions__question_id__get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Question Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cards": {
      "get": {
        "tags": [
          "cards"
        ],
        "summary": "List Cards",
        "operationId": "list_cards_api_cards_get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "due_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Filter only cards due for review now",
              "default": false,
              "title": "Due Only"
            },
            "description": "Filter only cards due for review now"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardOut"
                  },
                  "title": "Response List Cards Api Cards Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cards/{card_id}/review": {
      "post": {
        "tags": [
          "cards"
        ],
        "summary": "Review Card",
        "operationId": "review_card_api_cards__card_id__review_post",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Card Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardReviewIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/export/apkg": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Export Apkg",
        "operationId": "export_apkg_api_export_apkg_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/export/ankiconnect": {
      "post": {
        "tags": [
          "export"
        ],
        "summary": "Sync Ankiconnect",
        "operationId": "sync_ankiconnect_api_export_ankiconnect_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CardOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "mistake_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mistake Id"
          },
          "front": {
            "type": "string",
            "title": "Front"
          },
          "back": {
            "type": "string",
            "title": "Back"
          },
          "state": {
            "type": "integer",
            "title": "State"
          },
          "difficulty": {
            "type": "number",
            "title": "Difficulty"
          },
          "stability": {
            "type": "number",
            "title": "Stability"
          },
          "due_date": {
            "type": "string",
            "format": "date-time",
            "title": "Due Date"
          },
          "last_review": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Review"
          },
          "reps": {
            "type": "integer",
            "title": "Reps"
          },
          "lapses": {
            "type": "integer",
            "title": "Lapses"
          }
        },
        "type": "object",
        "required": [
          "id",
          "front",
          "back",
          "state",
          "difficulty",
          "stability",
          "due_date",
          "reps",
          "lapses"
        ],
        "title": "CardOut"
      },
      "CardReviewIn": {
        "properties": {
          "rating": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Rating",
            "description": "1=Again, 2=Hard, 3=Good, 4=Easy (FSRS Ratings)"
          }
        },
        "type": "object",
        "required": [
          "rating"
        ],
        "title": "CardReviewIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuestionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "model_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Answer"
          },
          "tips": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tips"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "difficulty",
          "title",
          "text"
        ],
        "title": "QuestionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}